Bump material lib to 1.13.0-alpha08 due to AOSP
AOSP 15 QPR2 uses version 1.13.0-alpha08 of the material library,
which somehow conflicts with our choice of 1.12.0 and causes a crash
on startup, making the app unusable. To work around this, we take the
"if you can't beat 'em, join 'em" approach and we vendor (bundle) the
same version for now.

Producing this commit involved the following steps:
1. Edit gradle/libs.versions.toml to change the version of material to
   1.13.0-alpha08.
2. Run ./gradlew generateBp.

(If you receive an error with generateBp, you may need to clear out
the ~/.gradle directory.)

Issue: calyxos#3057
Change-Id: Icb411d0cbe57be896825fe2c7190325dff36cc0d